requirements.txt
altair==4.2.0
numpy==1.23.3
pandas==1.4.4
plotly==5.10.0
scikit_learn==1.2.0
seaborn==0.12.1
streamlit==1.14.0
matplotlib==3.5.3
Pillow==8.4.0
latexify-py==0.4.2
tensorflow==2.7.0
tensorflow-addons
st-pages==0.5.0